test: colocate redaction policy tests with log module
